[CT-1355] Handle Suborders in End Blocker

+++ protocol/x/clob/keeper/msg_server_place_order_test.go
@@ -425,7 +425,7 @@ func TestPlaceOrder_Success(t *testing.T) {
 				twap_suborder, timestamp, found_suborder := ks.ClobKeeper.GetTwapTriggerPlacement(ctx, suborder.OrderId)
 				require.Equal(t, suborder.OrderId, twap_suborder)
 				require.True(t, found_suborder)
-				require.Equal(t, timestamp, uint64(ctx.BlockTime().Unix()))
+				require.Equal(t, timestamp, ctx.BlockTime().Unix())
 			} else {
 				_, found := ks.ClobKeeper.GetLongTermOrderPlacement(ctx, tc.StatefulOrderPlacement.GetOrderId())
 				require.True(t, found)
